Refactor package/plugin structure
This commit moves all python modules to sub directories and keeps an loader
plugin in root, only. It contains only public commands necessary for ST.

ST explicitly loads all python modules from root as plugins and looks for
EventListeners and Commands within them. This takes time and is therefore
not useful for unittests and other library-only modules such as ctags.py.

It may even cause harm, as importing ctags into ctagsplugin may cause
duplicated modules/globals being present.

Note: By using relative imports, this commit breaks compatibility with ST2.
