+++ .gitignore
@@ -1,2 +1,4 @@
-
+.mypy_cache
+.pytest_cache
+__pycache__
 *.pyc
