--- helpers/check_output.py
@@ -1,28 +1,0 @@
-"""
-Backport version of 'subprocess.check_output' for Python 2.6.x
-
-Based on source from here: https://gist.github.com/edufelipe/1027906
-"""
-
-import subprocess
-
-def check_output(*popenargs, **kwargs):
-    r"""
-    Run command with arguments and return its output as a byte string.
-
-    Backported from Python 2.7 as it's implemented as pure python on stdlib.
-
-    >>> check_output(['/usr/bin/python', '--version'])
-    Python 2.6.2
-    """
-    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
-    output, _ = process.communicate()
-    retcode = process.poll()
-    if retcode:
-        cmd = kwargs.get("args")
-        if cmd is None:
-            cmd = popenargs[0]
-            error = subprocess.CalledProcessError(retcode, cmd)
-            error.output = output
-            raise error
-    return output
